+++ src/mastodon_api/statuses/microsyntax/mentions.rs
@@ -12,9 +12,9 @@ use mitra_models::{
 
 use super::links::is_inside_code_block;
 
-// See also: ACTOR_ADDRESS_RE in mitra_federation::addresses
+// See also: USERNAME_RE in mitra_validators::profiles
 const MENTION_SEARCH_RE: &str = r"(?m)(?P<before>^|\s|>|[\(])@(?P<mention>[^\s<]+)";
-const MENTION_SEARCH_SECONDARY_RE: &str = r"^(?P<username>[\w\.-]+)(@(?P<hostname>[\w\.-]+\w))?(?P<after>[\.,:?!\)]?)$";
+const MENTION_SEARCH_SECONDARY_RE: &str = r"^(?P<username>[A-Za-z0-9\-\._]+)(@(?P<hostname>[\w\.-]+\w))?(?P<after>[\.,:?!\)]?)$";
 
 /// Finds everything that looks like a mention
 fn find_mentions(